Refactor schemas into a dedicated package
This change refactors the monolithic `schemas.py` file into a more organized package structure. Each domain now has its own schema file within the `server/domain/schemas/` directory.

This improves code organization and maintainability by grouping related schemas together. An `__init__.py` file has been added to the new package to ensure that all schemas can still be imported from the same top-level path as before, minimizing the impact on other parts of the codebase.
