+++ server/domain/schemas/owner.py
@@ -1,0 +1,13 @@
+from pydantic import BaseModel, ConfigDict
+from .enums import OrgType
+
+
+class OwnerStatsOut(BaseModel):
+    """Schema for an owner organization with crane fleet statistics."""
+
+    model_config = ConfigDict(from_attributes=True)
+
+    id: str
+    name: str
+    total_cranes: int
+    available_cranes: int
